backend/app/utils_mock.py
import requests
import os
from google.cloud import texttospeech
from dotenv import load_dotenv
from huggingface_hub import InferenceApi
from PIL import Image
from io import BytesIO
import base64

# Load environment variables
load_dotenv()

# Hugging Face API for text generation
huggingface_api = InferenceApi(repo_id="meta-llama/Meta-Llama-3-8B-Instruct", token=os.getenv('HUGGINGFACE_API_KEY'))

# Generate story part using Hugging Face model
def generate_story_part(genre, length, keywords, choice=None):

    return "The best story ever."

# Convert story text to audio using Google Cloud Text-to-Speech
def text_to_speech(text):
    return None

# Generate image based on the story using a text-to-image model
def generate_image(story_text):
    return None
