backend/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
requests==2.25.1
python-dotenv==0.15.0
google-cloud-texttospeech==2.4.0
huggingface-hub==0.1.2
Pillow==10.4.0
transformers
torch

